Changed to use device codename for ramdisk name

+++ runme_1unpack64.sh
@@ -3,11 +3,14 @@ if [ -d work ]; then
     rm -rf ./work
 fi
 
-if [ -e kernel*.elf ]; then
-    mv kernel*.elf kernel.elf
-fi
-if [ -e kernel*.ext4 ]; then
-    mv kernel*.ext4 kernel.elf
+if [ ! -e kernel.elf ]; then
+    if [ -e kernel*.elf ]; then
+        mv kernel*.elf kernel.elf
+    fi
+
+    if [ -e kernel*.ext4 ]; then
+        mv kernel*.ext4 kernel.elf
+    fi
 fi
 
 mkdir work
